requirements.txt
# VMC-Bot — dependencias base (Semana 1)

python-dotenv>=1.0.0
firecrawl-py>=1.0.0
pinecone>=5.0.0
anthropic>=0.39.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.32.0
